requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Jinja2
google-generativeai